clickable area of the back button, back from particular photocontest to the main page, hide extra icons, fixed upload posts, comments, removed navbar_buttons_render
