--- app/scripts/templates/content/restaurant_content.ejs
@@ -1,8 +1,0 @@
-<div id="cmntyex_restaurant" data-role="page" class="page">
-    <a href="#" class="more_button">More</a>
-    <div data-role="content" class="content">
-        <div class="cmntyex-gallery_placeholder">
-            <div class="cmntyex-content_placeholder"></div>
-        </div>
-    </div>
-</div>
