--- app/scripts/templates/partials/aboutUsButton.ejs
@@ -1,2 +1,0 @@
-<img src="styles/icons/aboutUs.png">
-<div class="navbar_buttons_text">Test1</div>
