--- app/scripts/templates/partials/contestButton.ejs
@@ -1,2 +1,0 @@
-<img src="styles/icons/Poll-contest1.png">
-<div class="navbar_buttons_text">Test1</div>
