--- app/scripts/templates/partials/promotionButton.ejs
@@ -1,2 +1,0 @@
-<img src="styles/icons/Promotion.png">
-<div class="navbar_buttons_text">Test1</div>
