--- app/scripts/templates/partials/restMenuButton.ejs
@@ -1,2 +1,0 @@
-<img src="styles/icons/Menu.png">
-<div class="navbar_buttons_text">Test1</div>
