--- app/scripts/templates/partials/signInButton.ejs
@@ -1,6 +1,0 @@
-<% if (isSignedIn) { %>
-    <img src="styles/icons/Sign_out1.png">
-<% } else { %>
-    <img src="styles/icons/Sign_in1.png">
-<% } %>
-<div class="navbar_buttons_text">Test1</div>
