+++ app/scripts/views/headers/navbarView.js
@@ -3,18 +3,21 @@
 'use strict';
 
 var Vent = require('../../Vent'),
-    RestMenuButton = require('../partials/restMenuButton'),
-    ContestButton = require('../partials/contestButton'),
-    AboutUsButton = require('../partials/aboutUsButton'),
-    SignInButton = require('../partials/signInButton'),
-    PromotionButton = require('../partials/promotionButton');
+    loader = require('../../loader'),
+    sessionActions = require('../../actions/sessionActions'),
+    userController = require('../../controllers/userController'),
+    promotionsController = require('../../controllers/promotionsController');
 
 var NavbarView = Backbone.View.extend({
 
     el: '#cmtyx_navbar',
 
     events: {
-        // 'click .menu_button_3': 'triggerContestsView'
+        'click .menu_button_1': 'openMenu',
+        'click .menu_button_2': 'openPromotion',
+        'click .menu_button_3': 'triggerContestsView',
+        'click .menu_button_4': 'triggerAboutUsView',
+        'click .menu_button_5': 'toggle'
     },
 
     initialize: function(options) {
@@ -22,19 +25,32 @@ var NavbarView = Backbone.View.extend({
         this.restaurant = options.restaurant;
         this.page = options.page;
 
+        this.user = sessionActions.getCurrentUser();
+
+        this.listenTo(Vent, 'login_success logout_success', this.render, this);
+
+        this.listenTo(this.parent, 'hide', this.remove, this);
+
         if (!this.page) {
             throw new Error('MapHeader::Expected a page');
         }
     },
 
-    render: function() {
-        this.renderRestMenuButton();
-        this.renderPromotionButton();
-        this.renderContestButton();
-        this.renderAboutUsButton();
-        this.renderSignInButton();
-
-        return this;
+    openPromotion: function(pid) {
+        loader.show('retrieving promotions');
+        promotionsController.fetchPromotionUUIDsBySasl(
+            this.restaurant.sa(),
+            this.restaurant.sl(),
+            this.page.user.getUID()
+        ).then(function(promotions) {
+            if(promotions.length < 1) {
+                loader.showFlashMessage('No promotions were found');
+            } else {
+                this.page.openSubview('promotions', promotions, {pid: pid, sasl: this.model});
+            }
+        }.bind(this), function () {
+            loader.showFlashMessage('error retrieving promotions');
+        });
     },
 
     triggerContestsView: function() {
@@ -43,36 +59,48 @@ var NavbarView = Backbone.View.extend({
         }.bind(this));
     },
 
-    renderRestMenuButton: function() {
-        this.$('.menu_button_1').html( new RestMenuButton({
-            parent: this.page,
-            model: this.restaurant
-        }).render().el);
+    triggerAboutUsView: function() {
+        if (this.model) {
+            Vent.trigger('viewChange', 'aboutUs', this.model.getUrlKey());
+        } else {
+            Vent.trigger('viewChange', 'aboutUs', [this.restaurant.sa(), this.restaurant.sl()]);
+        }
     },
 
-    renderContestButton: function() {
-        this.$('.menu_button_3').html( new ContestButton({
-            parent: this.page,
-            model: this.restaurant
-        }).render().el);
+    openMenu: function() {
+        this.page.openSubview('restaurantMenu', {}, this.restaurant.get('services'));
     },
 
-    renderAboutUsButton: function() {
-        this.$('.menu_button_4').html( new AboutUsButton({
-            parent: this.page,
-            model: this.restaurant
-        }).render().el);
+    confirmSignout: function () {
+        this.page.openSubview('confirmationPopup', {}, {
+            text: 'Are you sure you want to sign out?',
+            action: this.signout.bind(this)
+        });
+    },
+
+    signout: function() {
+        loader.show();
+        userController.logout(this.user.getUID()).then(function(){
+            loader.showFlashMessage( 'signed out' );
+        }, function(e){
+            loader.showFlashMessage(h().getErrorMessage(e, config.defaultErrorMsg));
+        });
+    },
+
+    signin: function() {
+        this.page.openSubview('signin', this.model);
+    },
+
+    toggle: function () {
+        if ( !this.user.getUID()) {
+            this.signin();
+        } else {
+            this.confirmSignout();
+        }
     },
 
     renderSignInButton: function() {
         this.$('.menu_button_5').html( new SignInButton({
-            parent: this.page,
-            model: this.restaurant
-        }).render().el);
-    },
-
-    renderPromotionButton: function() {
-        this.$('.menu_button_2').html( new PromotionButton({
             parent: this.page,
             model: this.restaurant
         }).render().el);
