+++ app/scripts/views/photoContestView.js
@@ -25,6 +25,7 @@ module.exports = PageLayout.extend({
     initialize: function(options) {
         options = options || {};
         this.sasl = options.sasl;
+        this.pageB = options.pageB;
         this.on('show', this.onShow, this);
     },
 
@@ -37,7 +38,7 @@ module.exports = PageLayout.extend({
     },
 
     triggerContestsView: function() {
-        Vent.trigger('viewChange', 'contests', this.sasl.getUrlKey(), { reverse: true });
+        Vent.trigger('viewChange', 'restaurant', this.sasl.getUrlKey(), { reverse: true })
     },
 
     renderPrizes: function () {
